finalização dos requisitos obrigatórios

+++ src/services/book.services.ts
@@ -1,12 +1,10 @@
 import { bookRepository } from "@/repositories/book.repositories";
-import { CreateBook } from "@/protocols/protocols";
 import { Book } from "@/protocols/protocols";
-import { number } from "joi";
-
-/* ************************************** REALIZAR AS VALIDAÇÕES ****************************** */
+import { errors } from "@/errors/errors";
 
 async function getById(id: number): Promise<Book> {
     const result = await bookRepository.getById(id);
+    if (!result) throw errors.notFound();
 
     return result;
 }
@@ -18,14 +16,23 @@ async function get(): Promise<Book[]> {
 }
 
 async function create(title: string, author: string, year: number, publisher: string): Promise<void> {
+    const result = await bookRepository.getByTitleAndYear(title, year);
+    if (result) throw errors.conflict();
+
     await bookRepository.create(title, author, year, publisher);
 }
 
 async function update(id: number, title: string, author: string, year: number, publisher: string): Promise<void> {
+    const result = await bookRepository.getById(id);
+    if (!result) throw errors.notFound();
+
     await bookRepository.update(id, title, author, year, publisher);
 }
 
 async function remove(id: number): Promise<void> {
+    const result = await bookRepository.getById(id);
+    if (!result) throw errors.notFound();
+
     await bookRepository.remove(id);
 }
